Bump to 1.0.0, separate cachign logic
