+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "lifx-lan-mcp",
   "description": "MCP server for Lifx LAN API integration",
-  "version": "0.0.1",
+  "version": "1.0.0",
   "main": "index.js",
   "author": "Simon Duchastel",
   "license": "MIT",
